Break up main
